修改mark类型为double

+++ src/main/java/org/example/server/controller/ScoreController.java
@@ -15,6 +15,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.nio.DoubleBuffer;
 import java.util.List;
 import java.util.concurrent.locks.ReentrantLock;
 
@@ -33,7 +34,7 @@ public class ScoreController {
     public Result insertScore(@Valid @RequestBody DataRequest dataRequest) {
         Integer student_id = dataRequest.getInteger("student_id");
         Integer course_id = dataRequest.getInteger("course_id");
-        Integer mark = dataRequest.getInteger("mark");
+        Double mark = dataRequest.getDouble("mark");
         return scoreService.insertScore(student_id, course_id, mark);
     }
 
@@ -56,7 +57,7 @@ public class ScoreController {
     public Result updateScore(@Valid @RequestBody DataRequest dataRequest) {
         Integer student_id = dataRequest.getInteger("student_id");
         Integer course_id = dataRequest.getInteger("course_id");
-        Integer mark = dataRequest.getInteger("mark");
+        Double mark = dataRequest.getDouble("mark");
         return scoreService.updateScoreAndRanking(student_id, course_id, mark);
     }
